+++ sdk/core/CHANGELOG.md
@@ -5,6 +5,8 @@
 - [minor] License Update
 - Android
   - [minor] Bundling `MoEFireBaseMessagingService` in Plugin's manifest to reduce integration steps for developers using FCM Push.
+- iOS
+  - [minor] Added no-code SDK file based initialization
 
 # 19-11-2025